feat: introduce asset pipelines
This commit refactors the build process to use asset pipelines, making
it more modular and extensible. The new architecture allows for building
new pipelines for different types of assets, not just HTML and Markdown.

The following changes are included:

- A new `pipelines` package has been added, which defines the `Asset`
and `Pipeline` interfaces.
- The build process has been updated to use the new asset pipelines.
- Pipelines for Markdown, HTML, and copying assets have been
implemented.
- A layout pipeline has been added to apply layouts to the processed
assets.
- Tests and benchmarks for the new asset pipelines have been added.

This new architecture will make it easier to add support for new asset
types in the future, such as CSS or JavaScript.